+++ 1.21.10/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.10.1-1.21.10] - 2025-11-18
+
+### Added
+
+- Add back optional integration for the [Reinforced Shulker Boxes](https://modrinth.com/mod/reinforced-shulker-boxes)
+  mod
+
+### Changed
+
+- All affected items are now split into separate built-in data packs, meaning you can choose if item interactions should
+  be enabled for shulker boxes, bundles, ender chests, containers, and maps separately
+- The built-in data packs containing shulker boxes, bundles, and ender chests stay enabled by default, while the
+  remaining items no longer are
+
 ## [v21.10.0-1.21.10] - 2025-10-08
 
 ### Changed